Rename module from number.rs to math.rs
