+++ src/flexbox.rs
@@ -6,8 +6,8 @@ use core::f32;
 use crate::forest::{Forest, NodeData};
 use crate::geometry::{Point, Rect, Size};
 use crate::layout::{Cache, Layout};
+use crate::math::MaybeMath;
 use crate::node::{MeasureFunc, NodeId};
-use crate::number::MaybeMath;
 use crate::style::{AlignContent, AlignSelf, Dimension, Display, FlexWrap, JustifyContent, PositionType};
 use crate::style::{FlexDirection, FlexboxLayout};
 use crate::sys::{abs, round, ChildrenVec, Vec};
